+++ data/saopigTaskList.txt
@@ -1,0 +1,4 @@
+D %&///&% 0 %&///&% read book 1 %&///&% 2024-02-12 12:00
+E %&///&% 0 %&///&% project meeting %&///&% 2024-01-02 12:00 %&///&% 2024-03-02 12:00
+D %&///&% 1 %&///&% return book %&///&% 2024-02-21 12:00
+E %&///&% 0 %&///&% project meeting2 %&///&% 2025-01-02 12:00 %&///&% 2025-03-02 12:00
