+++ src/main/java/saopig/task/Event.java
@@ -1,3 +1,5 @@
+package saopig.task;
+
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
