+++ src/main/java/saopig/task/Task.java
@@ -1,3 +1,5 @@
+package saopig.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
