+++ src/main/java/saopig/task/Todo.java
@@ -1,3 +1,5 @@
+package saopig.task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
